+++ OneDrive-Cloud-Player/OneDrive-Cloud-Player.csproj
@@ -123,9 +123,9 @@
     <Compile Include="Models\GraphData\CachedDriveItem.cs" />
     <Compile Include="Models\GraphData\OneDriveCache.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Services\Helpers\Caching\CacheHelper.cs" />
-    <Compile Include="Services\Helpers\API\GraphHelper.cs" />
-    <Compile Include="Services\Helpers\Authentication\GraphAuthHelper.cs" />
+    <Compile Include="Services\Helpers\CacheHelper.cs" />
+    <Compile Include="Services\Helpers\GraphHelper.cs" />
+    <Compile Include="Services\Helpers\GraphAuthHelper.cs" />
     <Compile Include="Services\Utilities\IO\JsonHandler.cs" />
     <Compile Include="ViewModels\LoginPageViewModel.cs" />
     <Compile Include="ViewModels\StartPageViewModel.cs" />
